rename gpu runtime files
